requirements.txt
pydantic~=1.8.2
fastapi~=0.68.2
python-socketio~=5.4.0